Refactor distributed planner into its own folder
