+++ src/distributed_planner/mod.rs
@@ -1,0 +1,7 @@
+mod distributed_physical_optimizer_rule;
+mod distributed_plan_error;
+mod network_boundary;
+
+pub use distributed_physical_optimizer_rule::DistributedPhysicalOptimizerRule;
+pub use distributed_plan_error::{DistributedPlanError, limit_tasks_err, non_distributable_err};
+pub use network_boundary::{InputStageInfo, NetworkBoundary, NetworkBoundaryExt};
