+++ src/execution_plans/distributed.rs
@@ -1,5 +1,5 @@
 use crate::channel_resolver_ext::get_distributed_channel_resolver;
-use crate::distributed_physical_optimizer_rule::NetworkBoundaryExt;
+use crate::distributed_planner::NetworkBoundaryExt;
 use crate::execution_plans::common::require_one_child;
 use crate::protobuf::DistributedCodec;
 use crate::stage::{ExecutionTask, Stage};
