+++ src/lib.rs
@@ -4,19 +4,19 @@ mod channel_resolver_ext;
 mod common;
 mod config_extension_ext;
 mod distributed_ext;
-mod distributed_physical_optimizer_rule;
 mod execution_plans;
 mod flight_service;
 mod metrics;
 mod stage;
 
+mod distributed_planner;
 mod protobuf;
 #[cfg(any(feature = "integration", test))]
 pub mod test_utils;
 
 pub use channel_resolver_ext::{BoxCloneSyncChannel, ChannelResolver};
 pub use distributed_ext::DistributedExt;
-pub use distributed_physical_optimizer_rule::{
+pub use distributed_planner::{
     DistributedPhysicalOptimizerRule, InputStageInfo, NetworkBoundary, NetworkBoundaryExt,
 };
 pub use execution_plans::{
